Made changes based on feedback from Wendell
